+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 ChangeLog
 =========
 
-2.2.1 (2024-09-24)
+2.2.2 (2024-09-24)
 ------------------
 
 - BUGFIX: really dispose video media file resources
 - IMPROVEMENT: show unselectable button on preset tab in grey
+- IMPROVEMENT: support non-looping videos
 - CLEANUP: remove light effects on background of Mask
 - UPDATE: update dependencies